grid: compute marching cubes in f64 with typed coordinates
Take a Voxels<f64> and keep vertices, normals and values in f64. Return vertices
as Coordinates and normals as Directions, and give spacing as a Coordinate of
lengths, so an Isosurface converts to a Tessellation without a copy. Parity with
scikit-image is now topological (identical faces) with vertices, normals and
values within a small tolerance. A normal that cancels to rounding noise is
returned as zero rather than normalized.

+++ src/geometry/mesh/tessellation/from/test.rs
@@ -33,7 +33,7 @@ fn isosurface_faces_outward() {
     let data = (0..216)
         .map(|i| {
             let [z, y, x] = [i / 36, i / 6 % 6, i % 6];
-            f32::from(u8::from(
+            f64::from(u8::from(
                 (2..4).contains(&z) && (2..4).contains(&y) && (2..4).contains(&x),
             ))
         })
